🚚(docs-package) move Divider in docs package
Move the Divider custom block to the common package.
It will be needed in multiple places, Docs app and
the y-provider server.
