+++ src/frontend/apps/impress/src/features/docs/doc-editor/components/custom-blocks/index.ts
@@ -1,2 +1,1 @@
 export * from './CalloutBlock';
-export * from './DividerBlock';
