+++ src/frontend/packages/common/src/custom-blocks/index.ts
@@ -1,0 +1,1 @@
+export * from './DividerBlock';
